Commit del ejercicio 4, terminado y corregido;

+++ src/java/modelo/Utilidades.java
@@ -26,9 +26,9 @@ public class Utilidades {
         ArrayList<Cuota> listaCuotas = new ArrayList<Cuota>();
         
         for (int i = 1; i <= num_cuotas; i++) {
-            int importeCuota = (int) (importeTotalPrestamo / num_cuotas);
-            int intereses = importeCuota - (importeCuota*(i-1) / num_cuotas);
-            int capital = importeCuota - intereses;
+            double importeCuota = (importeTotalPrestamo / num_cuotas);
+            double intereses = importeCuota - (importeCuota*(i-1) / num_cuotas);
+            double capital = importeCuota - intereses;
             Cuota c = new Cuota(i, importeCuota, capital,intereses);
             listaCuotas.add(c);
         }
